Peel some bits out of AIUtils to make it less unwieldy
